Initial implementation of soynode.
Supports compilation of soy files and loading of soy.js files.
Compilation is delegated to the SoyToJsSrcCompiler which is
executed via a system call to java.

Support for dynamic recompilation is available and temporary
files can be cleaned up after successful compilations.

This check-in includes a copy of Soy, downloaded on 22/05/12.