Adding i18n support - Fixes #2
I've been needing this functionality and saw that there was already a ticket open for it here, so this is implementing it.

We basically just need to add options for the --locales and --messageFilePathFormat options of SoyToJsSrcCompiler.jar, and use them.

I added a unit test for this and it works well :)